+++ dft2lnt/ASTValidator.h
@@ -62,7 +62,7 @@ public:
 		definedNodes.clear();
 		
 		// Go through all the AST nodes in a linear fashion
-		for(int i=0; i<ast->size(); ++i) {
+		for(size_t i=0; i<ast->size(); ++i) {
 			DFT::AST::ASTNode* node = ast->at(i);
 			assert(node);
 			switch(node->getType()) {
