+++ dft2lntc/dft2lntc.cpp
@@ -312,7 +312,7 @@ int main(int argc, char** argv) {
 			printHelp = true;
 		} else if (!strncmp(argv[argi], "-v", 2)) {
 			// -v
-			for (int i = 1; i < strlen(argv[argi]); i++) {
+			for (size_t i = 1; i < strlen(argv[argi]); i++) {
 				if (argv[argi][i] == 'v') {
 					++verbosity;
 				} else {
